requirements.txt
args==0.1.0
clint==0.4.1
pipeup==0.0.1
six==1.9.0
websocket-client==0.30.0
